Show command to fix Bad Tunables in HTML output
Hi,

attached patch adds another column into "Software settings in need of
tuning" table in HTML output. This column contains command needed to
tune the Tunables to fix the bad state.

Without that, it's not possible to find out what exactly is bad and how
the device (software setting) should be tuned to perform better.
Currently admins have to read the source code and then try to find out a
way how to fix bad tunables with common commands. With this patch it's
much more easier to do that.

Thanks for accepting,
Jan Kaluza