powertop: gracefully free (un)tunables
Gracefully free the vector of tunables and untunables on exit
from interactive mode and also when the report is generated.

This will also remove multiple entries of the same content
from the Chapter "Software Settings in need of Tuning" of the report
when iteration >= 2, for example when powertop is called as
# ./powertop --iteration=3 --html

The patch was verified with Valgrind:

valgrind --leak-check=full --log-file=powertop.log ./powertop

and

valgrind --leak-check=full --log-file=powertop.log ./powertop --iteration=3 --html

The patch removes the bigest chunk of leaked memory at exit from the Valgrind log:

==2925== 419,500 bytes in 25 blocks are definitely lost in loss record 125 of 130
==2925==    at 0x4025D54: operator new(unsigned int) (in /usr/lib/valgrind/vgpreload_memcheck-x86-linux.so)
==2925==    by 0x808D68D: add_runtime_tunables(char const*) (runtime.cpp:156)
==2925==    by 0x808889A: initialize_tuning() (tuning.cpp:71)
==2925==    by 0x805039E: main (main.cpp:391)

Note the size of one "tunable" which is 16780 bytes.

Signed-off-by: Márton Németh <nm127@freemail.hu>