Fix discharging detection
It seems there is an typo in the code that detects discharging.
There is another check in sysfs.cpp, so despite of this typo the
detection works correctly with sysfs backend. But the alternate
power_supply backend doesn't have this check, so if used, it
doesn't work correctly. I think the typo should be fixed or
the check should be removed from measurement.cpp (maybe
together with the power_supply implementation). The latter
solution is probably more clear.