Remove redundant inclusion of -O2 -g in powertop_CXXFLAGS
When building a program then program_CXXFLAGS replaces AM_CXXFLAGS and
CXXFLAGS is appended to it, this is the reason for the double -O2 -g options
when building the powertop binary.

This patch removes the extra -O2 -g and returns the control of the optimization
and debug flags to CXXFLAGS as one would expect.

Please note that the default value of CXXFLAGS is -O2 -g if the compiler
supports it.

Signed-off-by: Magnus Fromreide <magfr@lysator.liu.se>