Fix out-of-tree builds
Out-of-tree builds are a very useful feature which should be supported
by all programs which use GNU automake and the configure / make build.

For powertop, only a very small fix was needed to make it work.
Without that fix, powertop.css is not found when a build is not
started from the source tree, so csstoh fails and creates no css.h.

Signed-off-by: Stefan Weil <sw@weilnetz.de>