Make the "which C state line" logic better
the ARM guys complained that their human-readable C state names didn't have
numbers in them, and that as a result, the output is all messed up.
Using the "linux_name" instead is only a partial solution; it messes up the x86
side of the logic.

This patch fixes the logic to make the code use the human readable logic first,
but if there's no numbers there, fall back to the Linux name.

In addition, the patch allows callers to specify the line directly, overriding
both sets of logic.