Deduplication: move all instances of account_freq() to abstract_cpu::account_freq()
Trivial: unify identical in-class private definitions in the common base class.

The only override left is nhm_cpu::account_freq() which is almost identical but
ignores the "Turbo Mode" state.
Seems like a hidden bug to me, but will remove it in a separate commit so one can revert
it easily in case something goes wrong.

Signed-off-by: Ivan Shapovalov <intelfx100@gmail.com>