configure: use vertical lists
Horizontal lists has downside of being managed as a single entry in
version control, which makes commits that add and remove items difficult
to follow.  With vertical lists each entry is managed separately, which
is much nicer.

Additionally add project URL to autoconfig init, and pretty print macro
expressions to have indentation, and bracing that follows more or less in
K&R style.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>