add short options to usage, and reorder options in code
Earlier user experience was pretty unintuitive.  For example -h is often
help, and it generated following error message.

powertop: option requires an argument -- 'h'

This change re-assigned short option -h from --html to --help, and made
-r to be assigned with --html.  All other short vs long option pairs
remain untouched.

The option data in struct option, print_usage(), getopt_long() optstring,
and getopt_long switch() case, are reorderd be in same order in all four
places.  All options are in long-option alphabetical order, with
exception of --version and --help that are the last in the lists.

Some options where meant to be used as long-only options.  These are made
unreachable by making the option character to be out side of range what
user can type.

Finally if an unknown option is found exit the value is changed from zero
to one to indicate an error.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>