powertop: fix erroneous C state detection on ARM
When the kernel is configured with CONFIG_CPU_IDLE_MULTIPLE_DRIVERS,
the cpuidle directory contains a "driver" subdirectory that confuses
the current code parsing the C states on ARM platforms, ending up
with a C state entry corresponding to the actual driver name instead
of a proper C state entry.

This patch fixes the code by stopping the parsing if the files that
characterise the C states (ie usage) are not found in the respective:

/sys/devices/system/cpu/cpuX/cpuidle

subdirectory containing C states information.

Signed-off-by: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>
Tested-by: Kevin Hilman <khilman@linaro.org>