Reintroduced -d, --dump option, reports are by default written to STDOUT
The -d, --dump outputs plain text report, similarly how to old -d option
did. It takes one optional argument, the FILENAME. If there is no FILENAME
or it is '-' it outputs to STDOUT.

Similarly the CSV and HTML reporters now by default outputs to STDOUT,
which allows easy filtering of output through pipes. Previously their
FILENAME argument was mandatory, now it is optional.

Signed-off-by: Jaroslav Škarvada <jskarvad@redhat.com>