intel_cpus: frequency: clarify average clock speed
Depending on the sample time (-t sample_time_in_secs), on Intel CPUs the
average frequency is computed for the sample time. Therefore, let's
change the label to "Average" so as not to confuse the end-user.

Previously, the UI and report printed the "actual" frequency-- at least,
that's how it was labeled. This is not strictly true, as the actual cpu
frequency of the platform may change at any point during the sample time
for a variety of reasons.

Signed-off-by: Joe Konno <joe.konno@intel.com>