intel_cpus: no pstate Idle pct with intel_pstate
Do not display "percent Idle" in the Frequency Stats tab when the
intel_pstate kernel driver is loaded. This is because Idle time is
inferred through kernel PM trace events, which may be insufficient for
proper computation in certain scenarios. Therefore, best to display no
data than potentially inaccurate data.

End-users are encouraged to use the Idle Stats tab in the UI or report
to analyze their system's idle time. When the intel_pstate kernel driver
is loaded, the Frequency Stats tab will only present the end-user with
the average frequency of each logical core.

Signed-off-by: Joe Konno <joe.konno@intel.com>