configure.ac: version strings from git describe
There are two version strings we need during runtime: a long version
(for reports and --version), and a short version for the ncurses UI. Use
a utility script to generate those strings (version-long and
version-short) which will be cat'ed by configure and bundled when 'make
dist' is run.

Note: the version script will only run correctly within a clone of the
PowerTOP git repository. Consumers of a tarball, generated by 'make
dist', should not run autogen.sh or the version script. All necessary
configuration or build artifacts will be included in the tarball.

v2 (Magnus): Changed commit message to include note about intended
execution of the version script (and, by extension, autogen.sh).

Signed-off-by: Joe Konno <joe.konno@intel.com>