Doxyfile.in: add doxygen support
To assist PowerTOP developers, adding a Doxyfile for generating
developer documentation with Doxygen and Graphviz.

The Doxyfile will only generate HTML documentation by default, and it
will land in doc/html. Git will ignore both the autoconf-generated
Doxyfile and doc/html.

It makes no sense to bundle the Doxygen-generated documentation with
'make dist', so I didn't add support for it. It is intended solely for
PowerTOP developers, so I discourage distributions from including it in
their packages.

Signed-off-by: Joe Konno <joe.konno@intel.com>