src/devices/ahci.cpp: fix: Remove the humanname from ahci_data array
Device stats tab reports humanname for ahci and the purpose of having
a humanname is to identify ahci data clearly in the powertop report.
However, the printf() that is responsible for printing the humanname
spams the device stats tab and tunables tab for every auto refresh of
the window.

To fix this glitch, remove the humanname from the ahci_data array.

Signed-off-by: Gayatri Kammela <gayatri.kammela@intel.com>