src/tuning/tuning.cpp: fix: Update the size of tuned_data array
"Optimal Tuned Software Settings" is a section in the powertop report
shows all the settings that are tuned to GOOD. At present, this section
has multiple lines of whitespace at the end, which can confuse users as
if some data is missing.

Currently, the size of tuned_data array is set to size of all_tunables
and thus updating the size for the tuned_data array will remove the
whitespace.

Signed-off-by: Gayatri Kammela <gayatri.kammela@intel.com>