src/cpu/cpu.cpp: Bug fix: Increase the char buffer size in enumerate_cpus()
Function enumerate_cpus() fails to call handle_one_cpu() which is
responsible for populating supported cstates and pstates in a platform,
when get a line from /proc/cpuinfo is too big for the buffer's current
size of 1024.

In such scenario, since the cstates and pstates are zero, there
would be nothing to populate in "Idle stats" and Frequency stats", thus
leaving the tabs empty.

To fix up the issue, increase the buffer size to 4096.

Signed-off-by: Gayatri Kammela <gayatri.kammela@intel.com>