Fix root privileges check
PowerTOP must be run with root privileges, but `getuid()` checks the
user that the program is running as, and *not* whether the program is
running with root privileges. This is a subtle distinction but meant
that it was not possible to make the `powertop` binary setuid root.

Using `geteuid()` checks the *effective* user ID instead, which fixes
this bug.

Fixes #178.